requirements.txt
torch>=2.0.0
transformers>=4.30.0
opencv-python>=4.7.0
Pillow>=9.4.0
scikit-learn>=1.2.2
PyYAML>=6.0
tqdm==4.67.1
wandb==0.19.1
albumentations==1.4.22
evaluate==0.4.3
matplotlib==3.10.0
torchvision==0.20.1
notebook
ipython
ipywidgets